Moved implementation of ToJson to internal package
